Result class now supports std::string as the data type T.
The only downside is you must now use the Result<T>::Err() static factory method
to create errors, which adds a little bit of boilerplate. Some may argue that
the boilerplate makes the code harder to read, but it also makes the intent
more explicit (i.e., you know that a string really is an error condition because
the "Err()" clearly indicates it.

For convenience, you can still declare good results with Result<T>(value) constructor,
in addition to the new "Ok()" factory method. It seems pretty obvious that if you're
returning an actual piece of data rather than an error string, it means that you
have a good result, so no need for "Ok()" to mark it as good. But you can still do so
if you wish.